+++ src/main/java/nextstep/subway/line/application/PathPriceCalculator.java
@@ -4,14 +4,14 @@ import java.util.LinkedList;
 import java.util.List;
 import javax.annotation.PostConstruct;
 import nextstep.subway.line.domain.Line;
-import nextstep.subway.line.utils.AdultDiscountHelper;
-import nextstep.subway.line.utils.Between10And50PriceCalcHelper;
-import nextstep.subway.line.utils.ChildDiscountHelper;
-import nextstep.subway.line.utils.DiscountHelper;
-import nextstep.subway.line.utils.Over50PriceCalcHelper;
-import nextstep.subway.line.utils.PriceCalcHelper;
-import nextstep.subway.line.utils.StudentDiscountHelper;
-import nextstep.subway.line.utils.Under10PriceCalcHelper;
+import nextstep.subway.line.utils.Discount.AdultDiscountHelper;
+import nextstep.subway.line.utils.PriceCalculator.Between10And50PriceCalcHelper;
+import nextstep.subway.line.utils.Discount.ChildDiscountHelper;
+import nextstep.subway.line.utils.Discount.DiscountHelper;
+import nextstep.subway.line.utils.PriceCalculator.Over50PriceCalcHelper;
+import nextstep.subway.line.utils.PriceCalculator.PriceCalcHelper;
+import nextstep.subway.line.utils.Discount.StudentDiscountHelper;
+import nextstep.subway.line.utils.PriceCalculator.Under10PriceCalcHelper;
 import nextstep.subway.member.domain.Member;
 import org.springframework.stereotype.Component;
